cloudclient/aws: Move utilities from clusterinfo.go
