+++ pkg/cloudclient/aws/aws.go
@@ -21,7 +21,6 @@ import (
 	configv1 "github.com/openshift/api/config/v1"
 	cloudingressv1alpha1 "github.com/openshift/cloud-ingress-operator/pkg/apis/cloudingress/v1alpha1"
 	"github.com/openshift/cloud-ingress-operator/pkg/config"
-	utils "github.com/openshift/cloud-ingress-operator/pkg/controller/utils"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -96,7 +95,7 @@ func newClient(accessID, accessSecret, token, region string) (*Client, error) {
 
 // NewClient creates a new CloudClient for use with AWS.
 func NewClient(kclient client.Client) *Client {
-	region, err := utils.GetClusterRegion(kclient)
+	region, err := getClusterRegion(kclient)
 	if err != nil {
 		panic(fmt.Sprintf("Couldn't get cluster region %s", err.Error()))
 	}
