+++ pkg/cloudclient/aws/private.go
@@ -9,14 +9,18 @@ import (
 	"strings"
 	"time"
 
+	"gopkg.in/yaml.v2"
+
 	cloudingressv1alpha1 "github.com/openshift/cloud-ingress-operator/pkg/apis/cloudingress/v1alpha1"
 	utils "github.com/openshift/cloud-ingress-operator/pkg/controller/utils"
-	baseutils "github.com/openshift/cloud-ingress-operator/pkg/utils"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/types"
+	awsproviderapi "sigs.k8s.io/cluster-api-provider-aws/pkg/apis/awsproviderconfig/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/openshift/cloud-ingress-operator/pkg/config"
 	"github.com/openshift/cloud-ingress-operator/pkg/errors"
+	baseutils "github.com/openshift/cloud-ingress-operator/pkg/utils"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
@@ -44,6 +48,16 @@ type loadBalancerV2 struct {
 	loadBalancerName          string
 	scheme                    string
 	vpcID                     string
+}
+
+// installConfig represents the bare minimum requirement to get the AWS cluster region from the install-config
+// See https://bugzilla.redhat.com/show_bug.cgi?id=1814332
+type installConfig struct {
+	Platform struct {
+		AWS struct {
+			Region string `json:"region"`
+		} `json:"aws"`
+	} `json:"platform"`
 }
 
 // ensureAdminAPIDNS ensure the DNS record for the rh-api "admin API" for
@@ -115,7 +129,7 @@ func (c *Client) setDefaultAPIPublic(ctx context.Context, kclient client.Client,
 		return err
 	}
 	extNLBName := infrastructureName + "-ext"
-	subnets, err := utils.GetMasterNodeSubnets(kclient)
+	subnets, err := getMasterNodeSubnets(kclient)
 	if err != nil {
 		return err
 	}
@@ -173,6 +187,94 @@ func (c *Client) setDefaultAPIPublic(ctx context.Context, kclient client.Client,
 	}
 	// success
 	return nil
+}
+
+// getMasterNodeSubnets returns all the subnets for Machines with 'master' label.
+// return structure:
+// {
+//   public => subnetname,
+//   private => subnetname,
+// }
+//
+func getMasterNodeSubnets(kclient client.Client) (map[string]string, error) {
+	subnets := make(map[string]string)
+	machineList, err := baseutils.GetMasterMachines(kclient)
+	if err != nil {
+		return subnets, err
+	}
+	if len(machineList.Items) == 0 {
+		return subnets, fmt.Errorf("Did not find any master Machine objects")
+	}
+
+	// get the AZ from a Master object's providerSpec.
+	codec, err := awsproviderapi.NewCodec()
+
+	if err != nil {
+		return subnets, err
+	}
+
+	// Obtain the availability zone
+	awsconfig := &awsproviderapi.AWSMachineProviderConfig{}
+	err = codec.DecodeProviderSpec(&machineList.Items[0].Spec.ProviderSpec, awsconfig)
+	if err != nil {
+		return subnets, err
+	}
+
+	// Infra object gives us the Infrastructure name, which is the combination of
+	// cluster name and identifier.
+	infra, err := baseutils.GetInfrastructureObject(kclient)
+	if err != nil {
+		return subnets, err
+	}
+	subnets["public"] = fmt.Sprintf("%s-public-%s", infra.Status.InfrastructureName, awsconfig.Placement.AvailabilityZone)
+	subnets["private"] = fmt.Sprintf("%s-private-%s", infra.Status.InfrastructureName, awsconfig.Placement.AvailabilityZone)
+
+	return subnets, nil
+}
+
+//getClusterRegion returns the installed cluster's AWS region
+func getClusterRegion(kclient client.Client) (string, error) {
+	infra, err := baseutils.GetInfrastructureObject(kclient)
+	if err != nil {
+		return "", err
+	} else if infra.Status.PlatformStatus == nil {
+		// Try the deprecated configmap. See https://bugzilla.redhat.com/show_bug.cgi?id=1814332
+		return readClusterRegionFromConfigMap(kclient)
+	}
+	return infra.Status.PlatformStatus.AWS.Region, nil
+}
+
+func readClusterRegionFromConfigMap(kclient client.Client) (string, error) {
+	cm, err := getClusterConfigMap(kclient)
+	if err != nil {
+		return "", err
+	}
+	return parseClusterRegionFromConfigMap(cm)
+}
+
+func getClusterConfigMap(kclient client.Client) (*corev1.ConfigMap, error) {
+	cm := &corev1.ConfigMap{}
+	ns := types.NamespacedName{
+		Namespace: "kube-system",
+		Name:      "cluster-config-v1",
+	}
+	err := kclient.Get(context.TODO(), ns, cm)
+	return cm, err
+}
+
+func parseClusterRegionFromConfigMap(cm *corev1.ConfigMap) (string, error) {
+	if cm == nil || cm.Data == nil {
+		return "", fmt.Errorf("unexpected nil configmap or nil configmap Data")
+	}
+	data, ok := cm.Data["install-config"]
+	if !ok {
+		return "", fmt.Errorf("Missing install-config in configmap")
+	}
+	var ic installConfig
+	if err := yaml.Unmarshal([]byte(data), &ic); err != nil {
+		return "", fmt.Errorf("Invalid install-config: %v\njson:%s", err, data)
+	}
+	return ic.Platform.AWS.Region, nil
 }
 
 /* Helper functions below, sorted by AWS API type */
